[refactor]: GetBusSchedules 이름 변경
softeer5th#118
